+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Switched rate limiting to work correctly behind a reverse proxy.
 - 🤓 refactored a lot of the bookmarks database code for clarity and brevity.
+- 🤓 refactored title/description scraping to get rid of template and just use JSON.
 
 ## [0.1.4](https://github.com/anson-vandoren/pagepouch-rs/compare/v0.1.3...v0.1.4) - 2025-09-12